docs: revised reference doc task-tutorial.prereqs.md
In this commit the only change is the order of the list items, which isn't really important. I just wanted this page to show up in the changed files b/c I modified the structure of it to better fit the prereq section of the configure-redis-using-configmap.md file - BUT I made a mistake by merging that branch into main (to ensure it was successful). I reverted the merge but for a reason unknown to me the next commit didn't show the content changes for this page.